Moving all the tx-ops under TransactionOperation (can still statically import them)
