+++ crux-core/src/crux/api/v2/DeleteOperation.java
@@ -7,23 +7,19 @@ import java.util.Date;
 
 import static crux.api.v2.Util.keyword;
 
-public class DeleteOperation extends Operation {
+public class DeleteOperation extends TransactionOperation {
     private static final Keyword TX_DELETE = keyword("crux.tx/delete");
 
     private final Date validTime;
     private final CruxId deleteId;
 
-    private DeleteOperation(CruxId deleteId, Date validTime) {
+    DeleteOperation(CruxId deleteId, Date validTime) {
         this.deleteId = deleteId;
         this.validTime = validTime;
     }
 
     public DeleteOperation withValidTime(Date validTime) {
         return new DeleteOperation(deleteId, validTime);
-    }
-
-    public static DeleteOperation deleteOp(CruxId deleteId) {
-        return new DeleteOperation(deleteId, null);
     }
 
     @Override
