+++ crux-core/src/crux/api/v2/EvictOperation.java
@@ -5,17 +5,13 @@ import clojure.lang.PersistentVector;
 
 import static crux.api.v2.Util.keyword;
 
-public class EvictOperation extends Operation {
+public class EvictOperation extends TransactionOperation {
     private static final Keyword TX_EVICT = keyword("crux.tx/evict");
 
     private final CruxId evictId;
 
-    private EvictOperation(CruxId evictId) {
+    EvictOperation(CruxId evictId) {
         this.evictId = evictId;
-    }
-
-    public static EvictOperation evictOp(CruxId evictId) {
-        return new EvictOperation(evictId);
     }
 
     @Override
