--- crux-core/src/crux/api/v2/Operation.java
@@ -1,9 +1,0 @@
-package crux.api.v2;
-
-import clojure.lang.PersistentVector;
-
-public abstract class Operation {
-    Operation() { }
-
-    protected abstract PersistentVector toEdn();
-}
